+++ application/flicket/views/login.py
@@ -5,18 +5,20 @@
 
 from urllib.parse import urlparse, urljoin
 
-from flask import (flash,
-                   g,
-                   redirect,
-                   render_template,
-                   request,
-                   url_for)
-from flask_login import (current_user,
-                         login_user,
-                         logout_user)
-from flask_principal import (Identity,
-                             identity_changed)
+from flask import flash
+from flask import g
+from flask import redirect
+from flask import render_template
+from flask import request
+from flask import url_for
+from flask_login import current_user
+from flask_login import login_user
+from flask_login import logout_user
+from flask_principal import Identity
+from flask_principal import identity_changed
 from flask_babel import gettext
+from sqlalchemy import func
+from sqlalchemy import or_
 
 from application import app, lm, db, flicket_bp
 from application import __version__
@@ -99,7 +101,9 @@ def login():
     form = LogInForm()
 
     if form.validate_on_submit():
-        user = FlicketUser.query.filter_by(username=form.username.data).first()
+        user = FlicketUser.query.filter(
+            or_(FlicketUser.username == form.username.data,
+                func.lower(FlicketUser.email) == form.username.data.lower())).first()
         identity_changed.send(app, identity=Identity(user.id))
         login_user(user, remember=form.remember_me.data)
         # set the user token, authentication token is required for api use.